backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
Flask-Bcrypt
Flask-JWT-Extended
psycopg2-binary
python-dotenv
